+++ .eslintrc.js
@@ -30,8 +30,8 @@ module.exports = {
     'import/prefer-default-export': 0,
     'import/no-extraneous-dependencies': 0,
     'func-names': 0,
-    'camelcase': 1,
-    'quotes': 0,
+    camelcase: 1,
+    quotes: 0,
     'comma-dangle': 0,
     'no-useless-escape': 0,
     'no-tabs': 0,
