+++ src/callbacksReact/methods/advanced/preview.js
@@ -1,3 +1,20 @@
-// import { newLine } from '../../../domain/helper-methods';
+import { commonReplace } from '../../../domain/replace-wrapper3.0/commonReplace';
 
-// import { commonReplace } from '../../../domain/replace-wrapper3.0/commonReplace';
+function _previewText(text, content) {
+  const params = {
+    content,
+  };
+
+  const config = {
+    configurationMap: this.configurationMap,
+    params,
+    name: 'previewText',
+    debug: true,
+  };
+
+  // this.previewText = true;
+
+  const replaced = commonReplace(config);
+  return replaced;
+}
+export default _previewText;
