+++ src/constants/customTags.js
@@ -5,12 +5,17 @@ import os from 'os';
 const platform = os.platform();
 
 const newLine = platform === 'win32' ? '\r\n' : '\n';
-
+// those RegEx`es might gone after integration of front matter
 const REGEXP_SPONSORSHIP = /~(\[(.*?)\]){3}/g;
 const REGEXP_HTML_COMMENTS = /<!--(([\r\n]|.)*?)-->/g;
 // eslint-disable-next-line no-useless-escape
 const REGEXP_MEM = /\!\[(.*?)\]\[(.*?)\]\[(.*?)\]/g;
 const REGEXP_PREVIEW_TEXT = new RegExp(`#~(.*?)${newLine}`);
+//-------
+// TODO add that beforeMethod
+const regex_string = `\\[separator\\]`;
+const REGEXP_SEPARATOR = new RegExp(newLine + regex_string + newLine, 'g');
+//----------
 
 const REGEXP_HASH_TAG = new RegExp(
   // eslint-disable-next-line no-control-regex
@@ -23,10 +28,6 @@ const REGEXP_LINK_HTTPS = new RegExp('>https', 'g');
 
 const REGEXP_STR_BEGIN = '>https';
 const REGEXP_STR_END = 'f<';
-
-// TODO add that beforeMethod
-const regex_string = `\\[separator\\]`;
-const REGEXP_SEPARATOR = new RegExp(newLine + regex_string + newLine, 'g');
 
 export {
   REGEXP_SPONSORSHIP,
